move get_sync_token to base class
